Update the installation instructions
When creating the symbolic link, you must use the full path to the
executable (or at least a relative path that works). The example that
was given was a relative path that didn't work. The example is now
updated to use the 'realpath' utility to find the absolute path to
link to.

Signed-off-by: Joel Nider <joel@nider.org>