General cleanup
This commit contains several changes.

Verbose printing
The form [ $__verbose ] is now used throughout. It uses the [ builtin
to check for $__verbose != NULL. The global variable __verbose is unset
at the beginning of the script, and only set if the -v flag is passed
on the command line. The value of the variable doesn't really matter,
only the fact that it is set at all is what counts.

Function nesting level
The nesting level is increased from 10 to 20. I hit the limit when building
seL4 in its unoptimized form. Since the existing recipe is basically a
translation of the CMake build script, there are many many nested dependencies.
It is possible that 20 is too deep as a default, but I will leave it for
another day to be tuned properly. Even in the case of an infinitely recursive
dependency, 20 levels catches it quickly enough.

Version number
I added a version number __version="1.0" which can be printed from the
command line when _usage() is called (for example, with --help).

Internal functions
All internally defined functions start with an underscore.

Replaced all uses of 'eval' with 'local -n' for passing a variable containing
a name of another variable. This is used for automatically determining
dependencies of a target in a rule without explicitly knowing the target name.

Signed-off-by: Joel Nider <joel@nider.org>