Add built-in rules for c language
Rules for compiling and linking a program written in C are in the new
c_rules module. This can be included in any recipe with the new
'_language' function, which pulls in these rules at runtime.

The user has full control over which rules are included and which ones
are not. That helps keep the runtime to a minimum (since each rule must
be searched in turn for each target) while providing some default
functionality for common cases.

Signed-off-by: Joel Nider <joel@nider.org>