Add rules for cpp, cuda and ld
This adds language support for C++ and CUDA. Since they all produce .o
files, the rule for C language support has to be updated a bit as well
to differentiate between source file type.

The support for linking is only partially complete. It requires a bit
more thinking about how to include it, and how it should behave for
different linkers.

Signed-off-by: Joel Nider <joel@nider.org>