Filter target names with leading digit
Directory names can be used as target names. The rules for valid
directory names and valid bash variables are not the same, and this
is just one more instance of that problem. Any directory name that
starts with a digit must be prepended with an '_' to make it a valid
bash variable name.

Signed-off-by: Joel Nider <joel@nider.org>