Add better support for sub-recipes
A top-level recipe may call a recipe in a sub-directory for a specific
target. It can also build a directory as a target. It can now change
directories before reading the recipe (similar to make -C).

I'm still not convinced that I got this right but I need more test
cases to be sure.

Signed-off-by: Joel Nider <joel@nider.org>