Add support for compilation flags
When compiling source files, different sets of flags might be used
depending on the source file. This gives the flexibility of using
a global associative array 'flags' to hold the set of flags on a
per-target basis. If the array is not set for a particular target,
a generic variable (CFLAGS for .c files) will be used.

Signed-off-by: Joel Nider <joel@nider.org>