Use dependency list for build target
When compiling source files, it may not be obvious how to derive the
source file name from the target (.o) file name. The source file can
be in a different directory, or the file name might not be related
at all.

Dependencies of each source file are already supported by defining
an array of file names, with the name of the array as a filtered
name of the target. The first file in the dependency array is now
used by the compilation rule as the source file name. If there are
no dependencies, the rule will guess the source name using the
same logic as before this change.

Signed-off-by: Joel Nider <joel@nider.org>