Use 'objects' array for linker
Linking rules are slightly different than compilation rules, because
there can be many input files that need to be linked together to
create a single output file, rather than a single input file.
The dependency list was used to supply the input files to the linker
but this is not correct. Linking targets may have dependencies that
are not meant to be linked for example, linker scripts. To better
support linking, the objects that are meant to be linked are kept in
a separate array called 'objects'. It is likely that all objects
should also be dependencies, but not vice-versa.

Signed-off-by: Joel Nider <joel@nider.org>