Use global LDFLAGS for linking
If there are no target-specific flags passed to the linker rule, use
the global environment variable LDFLAGS instead.

Also, pass a null string by default if there are no objects defined
for the link target. This probably warrants a warning message, since
it is unlikely that someone will want to link an empty file (but not
impossible).

Signed-off-by: Joel Nider <joel@nider.org>