linker: use 'objects' array
The user can create an array called 'objects' to explicitly specify files
that are to be linked. This is necessary to separate the dependencies of
the linked file from the input objects. The dependencies are generally a
superset of the input objects and can include other files such as linker
scripts that must be treated differently than input objects.

Signed-off-by: Joel Nider <joel@nider.org>