object file names are case insensitive
Generally output objects are named with the .o suffix but not necessarily.
This allows output objects to be named with an uppercase .O instead. This
increases flexibility a bit, which helps with files coming from operating
systems that are case-insensitive and may use only uppercase letters.

It is also important to verify that the output file name is of the correct
type. Otherwise, the rule can be fooled into building the wrong kind of
file, if its first dependency happens to be a supported file type.

Signed-off-by: Joel Nider <joel@nider.org>